Lib tested, not fully optimized

+++ lib_dev/MAX11615_libDev/lib/MAX1161X/MAX1161X.h
@@ -56,9 +56,9 @@ MAX1161X_INTF_RET_TYPE max1161x_setConfigByte(uint8_t config, struct MAX1161X_De
 
 MAX1161X_INTF_RET_TYPE max1161x_setExternalVRef(float vref, struct MAX1161X_Dev_t *dev);
 
-MAX1161X_INTF_RET_TYPE max1161x_readADC_singleEnded(uint8_t channel, uint16_t *data, struct MAX1161X_Dev_t *dev);
+MAX1161X_INTF_RET_TYPE max1161x_readADC_singleEnded(uint8_t channel, float *data, struct MAX1161X_Dev_t *dev);
 
-MAX1161X_INTF_RET_TYPE max1161x_readADC_differential(uint8_t channel, uint16_t *data, struct MAX1161X_Dev_t *dev);
+MAX1161X_INTF_RET_TYPE max1161x_readADC_differential(uint8_t channel, float *data, struct MAX1161X_Dev_t *dev);
 
 #ifdef __cplusplus
 }
